requirements.txt
torch
torchvision
torchaudio
numpy
pandas
matplotlib
scikit-learn
tqdm
pillow
opencv-python
streamlit
